src/requirements.txt
requests
beautifulsoup4
pandas
numpy
yfinance
schedule
python-dotenv